refactor(search): consolidate template search into search module
