+++ src/modules/search/application/store/cachedSearchCacheStore.ts
