+++ src/modules/search/application/usecases/templateSearchLogic.ts
